Rename module `thunks` to `effects`
This commit renames the module `thunks` to `effects`. "Thunk" is a very general term. In contrast, "effect" can lean on ngrx/effects library.
