+++ src/__tests__/components/PreviousQuestion.test.tsx
@@ -6,9 +6,9 @@ import thunk from "redux-thunk";
 
 import { ASK_TO_REFOCUS } from "../../actions";
 import { PreviousQuestion } from "../../components/PreviousQuestion";
-import * as thunks from "../../thunks";
+import * as effects from "../../effects";
 
-jest.mock("../../thunks");
+jest.mock("../../effects");
 
 it("provokes the state change.", async () => {
   const store = configureMockFactory([thunk])();
@@ -18,7 +18,7 @@ it("provokes the state change.", async () => {
 
   const dummy = { hello: 1 };
 
-  (thunks.previousQuestion as any).mockResolvedValue(dummy);
+  (effects.previousQuestion as any).mockResolvedValue(dummy);
 
   const rendered = render(
     <Provider store={store}>
