+++ src/__tests__/reducers.test.ts
@@ -1,6 +1,6 @@
+import * as effects from "../effects";
 import { questionBank } from "../QuestionBank";
 import * as storeFactory from "../storeFactory";
-import { nextQuestion, previousQuestion } from "../thunks";
 
 const deps = { questionBank };
 
@@ -12,7 +12,7 @@ it("initializes the current question to the first question.", () => {
 
 it("selects the second question on going forward upon initialization.", () => {
   const store = storeFactory.produce(deps);
-  store.dispatch(nextQuestion() as any);
+  store.dispatch(effects.nextQuestion() as any);
 
   if (deps.questionBank.questions.length > 1) {
     expect(store.getState().currentQuestion).toBe(questionBank.questions[1].id);
@@ -21,7 +21,7 @@ it("selects the second question on going forward upon initialization.", () => {
 
 it("selects the last question on going backwards upon initialization.", () => {
   const store = storeFactory.produce(deps);
-  store.dispatch(previousQuestion() as any);
+  store.dispatch(effects.previousQuestion() as any);
 
   if (deps.questionBank.questions.length > 0) {
     expect(store.getState().currentQuestion).toBe(
