+++ src/components/PreviousQuestion.tsx
@@ -3,16 +3,16 @@ import ArrowLeft from "@material-ui/icons/ArrowLeft";
 import * as React from "react";
 import { useDispatch } from "react-redux";
 
-import { askToRefocus } from "../actions";
-import { previousQuestion } from "../thunks";
+import * as actions from "../actions";
+import * as effects from "../effects";
 
 export function PreviousQuestion() {
   const dispatch = useDispatch();
   return (
     <IconButton
       onClick={() => {
-        dispatch(previousQuestion());
-        dispatch(askToRefocus());
+        dispatch(effects.previousQuestion());
+        dispatch(actions.askToRefocus());
       }}
       data-testid="previousQuestion"
     >
